backend/requirements.txt
Django==5.0.2
django-import-export==3.3.6
django-admin-rangefilter==0.12.0
django-admin-list-filter-dropdown==1.0.2
python-dateutil==2.8.2
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2 